disabling old code and testing encoders
