+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/inactiveTestCode/AutonPlaceParking/AutonRedBackPlacePark.java
@@ -1,12 +1,14 @@
-package org.firstinspires.ftc.teamcode.AutonPlaceParking;
+package org.firstinspires.ftc.teamcode.inactiveTestCode.AutonPlaceParking;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 
-@Autonomous(name = "Auton Blue Front Place Park")
-public class AutonBlueFrontPlacePark extends LinearOpMode {
+@Autonomous(name = "Auton Red Back Place Park")
+@Disabled
+public class AutonRedBackPlacePark extends LinearOpMode {
 
     private DcMotor leftFrontDrive   = null;  //  Used to control the left front drive wheel
     private DcMotor rightFrontDrive  = null;  //  Used to control the right front drive wheel
@@ -53,19 +55,19 @@ public class AutonBlueFrontPlacePark extends LinearOpMode {
             rightFrontDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
             rightBackDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
 
-            rightFrontDrive.setTargetPosition(-250);
+            rightFrontDrive.setTargetPosition(-1480);
             rightFrontDrive.setPower(0.5);
             rightFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
-            leftBackDrive.setTargetPosition(-250);
+            leftBackDrive.setTargetPosition(-1480);
             leftBackDrive.setPower(0.5);
             leftBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
-            leftFrontDrive.setTargetPosition(250);
+            leftFrontDrive.setTargetPosition(1480);
             leftFrontDrive.setPower(0.5);
             leftFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
-            rightBackDrive.setTargetPosition(250);
+            rightBackDrive.setTargetPosition(1480);
             rightBackDrive.setPower(0.5);
             rightBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
@@ -90,10 +92,10 @@ public class AutonBlueFrontPlacePark extends LinearOpMode {
             leftBackDrive.setPower(0);
             }
 
-            int leftFrontDriveNecessaryTicks = calculateTicksForLateralMovement(30); //2000
-            int rightFrontDriveNecessaryTicks = calculateTicksForLateralMovement(30);
-            int leftBackDriveNecessaryTicks = calculateTicksForLateralMovement(30);
-            int rightBackDriveNecessaryTicks = calculateTicksForLateralMovement(30);
+            int leftFrontDriveNecessaryTicks = calculateTicksForLateralMovement(85); //2000
+            int rightFrontDriveNecessaryTicks = calculateTicksForLateralMovement(85);
+            int leftBackDriveNecessaryTicks = calculateTicksForLateralMovement(85);
+            int rightBackDriveNecessaryTicks = calculateTicksForLateralMovement(85);
 
             int leftFrontDriveTargetTicks =  leftFrontDriveNecessaryTicks;
             int rightFrontDriveTargetTicks = rightFrontDriveNecessaryTicks;
@@ -105,19 +107,19 @@ public class AutonBlueFrontPlacePark extends LinearOpMode {
             leftFrontDrive.setDirection(DcMotorSimple.Direction.REVERSE);
             leftBackDrive.setDirection(DcMotorSimple.Direction.REVERSE);
 
-            leftFrontDrive.setTargetPosition(leftFrontDrive.getCurrentPosition() + leftFrontDriveTargetTicks);
+            leftFrontDrive.setTargetPosition(leftFrontDrive.getCurrentPosition() - leftFrontDriveTargetTicks);
             leftFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
             leftFrontDrive.setPower(0.5);
 
-            rightFrontDrive.setTargetPosition(rightFrontDrive.getCurrentPosition() + rightFrontDriveTargetTicks);
+            rightFrontDrive.setTargetPosition(rightFrontDrive.getCurrentPosition() - rightFrontDriveTargetTicks);
             rightFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
             rightFrontDrive.setPower(0.5);
 
-            leftBackDrive.setTargetPosition(leftBackDrive.getCurrentPosition() + leftBackDriveTargetTicks);
+            leftBackDrive.setTargetPosition(leftBackDrive.getCurrentPosition() - leftBackDriveTargetTicks);
             leftBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
             leftBackDrive.setPower(0.5);
 
-            rightBackDrive.setTargetPosition(rightBackDrive.getCurrentPosition() + rightBackDriveTargetTicks);
+            rightBackDrive.setTargetPosition(rightBackDrive.getCurrentPosition() - rightBackDriveTargetTicks);
             rightBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
             rightBackDrive.setPower(0.5);
