+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/inactiveTestCode/AutonPlaceParking/AutonRedBackwardsPlacePark.java
@@ -1,12 +1,17 @@
-package org.firstinspires.ftc.teamcode.AutonPlaceParking;
+package org.firstinspires.ftc.teamcode.inactiveTestCode.AutonPlaceParking;
 
+import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
+import com.qualcomm.robotcore.hardware.IMU;
+import com.qualcomm.robotcore.hardware.Servo;
 
-@Autonomous(name = "Auton Red Back Place Park")
-public class AutonRedBackPlacePark extends LinearOpMode {
+@Autonomous(name = "Auton Red Backwards Place Park")
+@Disabled
+public class AutonRedBackwardsPlacePark extends LinearOpMode {
 
     private DcMotor leftFrontDrive   = null;  //  Used to control the left front drive wheel
     private DcMotor rightFrontDrive  = null;  //  Used to control the right front drive wheel
@@ -43,6 +48,27 @@ public class AutonRedBackPlacePark extends LinearOpMode {
         leftBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
         rightBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
 
+        DcMotor arm = hardwareMap.dcMotor.get("arm");
+        Servo clawLeft = hardwareMap.servo.get("clawLeft");
+        Servo clawRight = hardwareMap.servo.get("clawRight");
+        Servo axle = hardwareMap.servo.get("axle");
+
+        arm.setDirection(DcMotorSimple.Direction.REVERSE);
+        clawLeft.setDirection(Servo.Direction.FORWARD);
+        clawRight.setDirection(Servo.Direction.REVERSE);
+
+        arm.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        arm.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+
+        clawLeft.scaleRange(0, 1);
+        clawRight.scaleRange(0, 1);
+
+        axle.scaleRange(0, 1);
+
+        int liftTargetPosition = 0;
+        double openClaw = 0.4; //increase to open more
+        double closeClaw = 0.2;
+
         waitForStart();
 
         if (opModeIsActive()) {
@@ -53,19 +79,23 @@ public class AutonRedBackPlacePark extends LinearOpMode {
             rightFrontDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
             rightBackDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
 
-            rightFrontDrive.setTargetPosition(-1480);
+            clawLeft.setPosition(closeClaw);
+            clawRight.setPosition(closeClaw);
+            axle.setPosition(0.8);
+
+            rightFrontDrive.setTargetPosition(-1450);
             rightFrontDrive.setPower(0.5);
             rightFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
-            leftBackDrive.setTargetPosition(-1480);
+            leftBackDrive.setTargetPosition(-1450);
             leftBackDrive.setPower(0.5);
             leftBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
-            leftFrontDrive.setTargetPosition(1480);
+            leftFrontDrive.setTargetPosition(1450);
             leftFrontDrive.setPower(0.5);
             leftFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
-            rightBackDrive.setTargetPosition(1480);
+            rightBackDrive.setTargetPosition(1450);
             rightBackDrive.setPower(0.5);
             rightBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
@@ -90,10 +120,46 @@ public class AutonRedBackPlacePark extends LinearOpMode {
             leftBackDrive.setPower(0);
             }
 
-            int leftFrontDriveNecessaryTicks = calculateTicksForLateralMovement(85); //2000
-            int rightFrontDriveNecessaryTicks = calculateTicksForLateralMovement(85);
-            int leftBackDriveNecessaryTicks = calculateTicksForLateralMovement(85);
-            int rightBackDriveNecessaryTicks = calculateTicksForLateralMovement(85);
+            rightFrontDrive.setTargetPosition(leftFrontDrive.getCurrentPosition() - 10);
+            rightFrontDrive.setPower(0.5);
+            rightFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+
+            leftBackDrive.setTargetPosition(leftFrontDrive.getCurrentPosition() + 10);
+            leftBackDrive.setPower(0.5);
+            leftBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+
+            leftFrontDrive.setTargetPosition(leftFrontDrive.getCurrentPosition() + 10);
+            leftFrontDrive.setPower(0.5);
+            leftFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+
+            rightBackDrive.setTargetPosition(leftFrontDrive.getCurrentPosition() - 10);
+            rightBackDrive.setPower(0.5);
+            rightBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+
+            telemetry.addData("Front Left Target Position",leftFrontDrive.getTargetPosition());
+            telemetry.addData("Front Left Current Position Position",leftFrontDrive.getCurrentPosition());
+
+            telemetry.addData("Front Right Target Position",rightFrontDrive.getTargetPosition());
+            telemetry.addData("Front Right Current Position Position",rightFrontDrive.getCurrentPosition());
+
+            telemetry.addData("Back Left Target Position",leftBackDrive.getTargetPosition());
+            telemetry.addData("Back Left Current Position Position",leftBackDrive.getCurrentPosition());
+
+            telemetry.addData("Back Right Target Position",rightBackDrive.getTargetPosition());
+            telemetry.addData("Back Right Current Position Position",rightBackDrive.getCurrentPosition());
+
+            telemetry.update();
+
+            while (leftFrontDrive.isBusy() || rightFrontDrive.isBusy() || leftBackDrive.isBusy() || rightBackDrive.isBusy()) {}
+            rightFrontDrive.setPower(0);
+            leftFrontDrive.setPower(0);
+            rightBackDrive.setPower(0);
+            leftBackDrive.setPower(0);
+
+            int leftFrontDriveNecessaryTicks = calculateTicksForLateralMovement(83); //2000
+            int rightFrontDriveNecessaryTicks = calculateTicksForLateralMovement(83);
+            int leftBackDriveNecessaryTicks = calculateTicksForLateralMovement(83);
+            int rightBackDriveNecessaryTicks = calculateTicksForLateralMovement(83);
 
             int leftFrontDriveTargetTicks =  leftFrontDriveNecessaryTicks;
             int rightFrontDriveTargetTicks = rightFrontDriveNecessaryTicks;
@@ -142,6 +208,31 @@ public class AutonRedBackPlacePark extends LinearOpMode {
             rightBackDrive.setPower(0);
             leftBackDrive.setPower(0);
 
+            for (int liftPos = 0; liftPos <= 500; liftPos = liftPos + 5) {
+                liftTargetPosition = liftPos;
+                arm.setTargetPosition(liftTargetPosition);
+                arm.setPower(0.7);
+                arm.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+                sleep(15);
+            }
+
+            clawLeft.setPosition(openClaw);
+            clawRight.setPosition(openClaw);
+
+            sleep(200);
+            axle.setPosition(0);
+            sleep(500);
+
+            liftTargetPosition = 10;
+            arm.setTargetPosition(liftTargetPosition);
+            arm.setPower(0.05);
+            arm.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+
+            sleep(500);
+            axle.setPosition(0.8);
+
+            while (arm.isBusy()) {}
+            arm.setPower(0);
 
     }
