+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/inactiveTestCode/AutonPlaceParking/AutonRedFrontPlacePark.java
@@ -1,15 +1,14 @@
-package org.firstinspires.ftc.teamcode.AutonPlaceParking;
+package org.firstinspires.ftc.teamcode.inactiveTestCode.AutonPlaceParking;
 
-import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.robotcore.hardware.IMU;
-import com.qualcomm.robotcore.hardware.Servo;
 
-@Autonomous(name = "Auton Red Backwards Place Park")
-public class AutonRedBackwardsPlacePark extends LinearOpMode {
+@Autonomous(name = "Auton Red Front Place Park")
+@Disabled
+public class AutonRedFrontPlacePark extends LinearOpMode {
 
     private DcMotor leftFrontDrive   = null;  //  Used to control the left front drive wheel
     private DcMotor rightFrontDrive  = null;  //  Used to control the right front drive wheel
@@ -46,27 +45,6 @@ public class AutonRedBackwardsPlacePark extends LinearOpMode {
         leftBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
         rightBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
 
-        DcMotor arm = hardwareMap.dcMotor.get("arm");
-        Servo clawLeft = hardwareMap.servo.get("clawLeft");
-        Servo clawRight = hardwareMap.servo.get("clawRight");
-        Servo axle = hardwareMap.servo.get("axle");
-
-        arm.setDirection(DcMotorSimple.Direction.REVERSE);
-        clawLeft.setDirection(Servo.Direction.FORWARD);
-        clawRight.setDirection(Servo.Direction.REVERSE);
-
-        arm.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        arm.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-
-        clawLeft.scaleRange(0, 1);
-        clawRight.scaleRange(0, 1);
-
-        axle.scaleRange(0, 1);
-
-        int liftTargetPosition = 0;
-        double openClaw = 0.4; //increase to open more
-        double closeClaw = 0.2;
-
         waitForStart();
 
         if (opModeIsActive()) {
@@ -77,23 +55,19 @@ public class AutonRedBackwardsPlacePark extends LinearOpMode {
             rightFrontDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
             rightBackDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
 
-            clawLeft.setPosition(closeClaw);
-            clawRight.setPosition(closeClaw);
-            axle.setPosition(0.8);
-
-            rightFrontDrive.setTargetPosition(-1450);
+            rightFrontDrive.setTargetPosition(250);
             rightFrontDrive.setPower(0.5);
             rightFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
-            leftBackDrive.setTargetPosition(-1450);
+            leftBackDrive.setTargetPosition(250);
             leftBackDrive.setPower(0.5);
             leftBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
-            leftFrontDrive.setTargetPosition(1450);
+            leftFrontDrive.setTargetPosition(-250);
             leftFrontDrive.setPower(0.5);
             leftFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
-            rightBackDrive.setTargetPosition(1450);
+            rightBackDrive.setTargetPosition(-250);
             rightBackDrive.setPower(0.5);
             rightBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
@@ -118,46 +92,10 @@ public class AutonRedBackwardsPlacePark extends LinearOpMode {
             leftBackDrive.setPower(0);
             }
 
-            rightFrontDrive.setTargetPosition(leftFrontDrive.getCurrentPosition() - 10);
-            rightFrontDrive.setPower(0.5);
-            rightFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-
-            leftBackDrive.setTargetPosition(leftFrontDrive.getCurrentPosition() + 10);
-            leftBackDrive.setPower(0.5);
-            leftBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-
-            leftFrontDrive.setTargetPosition(leftFrontDrive.getCurrentPosition() + 10);
-            leftFrontDrive.setPower(0.5);
-            leftFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-
-            rightBackDrive.setTargetPosition(leftFrontDrive.getCurrentPosition() - 10);
-            rightBackDrive.setPower(0.5);
-            rightBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-
-            telemetry.addData("Front Left Target Position",leftFrontDrive.getTargetPosition());
-            telemetry.addData("Front Left Current Position Position",leftFrontDrive.getCurrentPosition());
-
-            telemetry.addData("Front Right Target Position",rightFrontDrive.getTargetPosition());
-            telemetry.addData("Front Right Current Position Position",rightFrontDrive.getCurrentPosition());
-
-            telemetry.addData("Back Left Target Position",leftBackDrive.getTargetPosition());
-            telemetry.addData("Back Left Current Position Position",leftBackDrive.getCurrentPosition());
-
-            telemetry.addData("Back Right Target Position",rightBackDrive.getTargetPosition());
-            telemetry.addData("Back Right Current Position Position",rightBackDrive.getCurrentPosition());
-
-            telemetry.update();
-
-            while (leftFrontDrive.isBusy() || rightFrontDrive.isBusy() || leftBackDrive.isBusy() || rightBackDrive.isBusy()) {}
-            rightFrontDrive.setPower(0);
-            leftFrontDrive.setPower(0);
-            rightBackDrive.setPower(0);
-            leftBackDrive.setPower(0);
-
-            int leftFrontDriveNecessaryTicks = calculateTicksForLateralMovement(83); //2000
-            int rightFrontDriveNecessaryTicks = calculateTicksForLateralMovement(83);
-            int leftBackDriveNecessaryTicks = calculateTicksForLateralMovement(83);
-            int rightBackDriveNecessaryTicks = calculateTicksForLateralMovement(83);
+            int leftFrontDriveNecessaryTicks = calculateTicksForLateralMovement(30); //2000
+            int rightFrontDriveNecessaryTicks = calculateTicksForLateralMovement(30);
+            int leftBackDriveNecessaryTicks = calculateTicksForLateralMovement(30);
+            int rightBackDriveNecessaryTicks = calculateTicksForLateralMovement(30);
 
             int leftFrontDriveTargetTicks =  leftFrontDriveNecessaryTicks;
             int rightFrontDriveTargetTicks = rightFrontDriveNecessaryTicks;
@@ -169,19 +107,19 @@ public class AutonRedBackwardsPlacePark extends LinearOpMode {
             leftFrontDrive.setDirection(DcMotorSimple.Direction.REVERSE);
             leftBackDrive.setDirection(DcMotorSimple.Direction.REVERSE);
 
-            leftFrontDrive.setTargetPosition(leftFrontDrive.getCurrentPosition() - leftFrontDriveTargetTicks);
+            leftFrontDrive.setTargetPosition(leftFrontDrive.getCurrentPosition() + leftFrontDriveTargetTicks);
             leftFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
             leftFrontDrive.setPower(0.5);
 
-            rightFrontDrive.setTargetPosition(rightFrontDrive.getCurrentPosition() - rightFrontDriveTargetTicks);
+            rightFrontDrive.setTargetPosition(rightFrontDrive.getCurrentPosition() + rightFrontDriveTargetTicks);
             rightFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
             rightFrontDrive.setPower(0.5);
 
-            leftBackDrive.setTargetPosition(leftBackDrive.getCurrentPosition() - leftBackDriveTargetTicks);
+            leftBackDrive.setTargetPosition(leftBackDrive.getCurrentPosition() + leftBackDriveTargetTicks);
             leftBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
             leftBackDrive.setPower(0.5);
 
-            rightBackDrive.setTargetPosition(rightBackDrive.getCurrentPosition() - rightBackDriveTargetTicks);
+            rightBackDrive.setTargetPosition(rightBackDrive.getCurrentPosition() + rightBackDriveTargetTicks);
             rightBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
             rightBackDrive.setPower(0.5);
 
@@ -206,31 +144,6 @@ public class AutonRedBackwardsPlacePark extends LinearOpMode {
             rightBackDrive.setPower(0);
             leftBackDrive.setPower(0);
 
-            for (int liftPos = 0; liftPos <= 500; liftPos = liftPos + 5) {
-                liftTargetPosition = liftPos;
-                arm.setTargetPosition(liftTargetPosition);
-                arm.setPower(0.7);
-                arm.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-                sleep(15);
-            }
-
-            clawLeft.setPosition(openClaw);
-            clawRight.setPosition(openClaw);
-
-            sleep(200);
-            axle.setPosition(0);
-            sleep(500);
-
-            liftTargetPosition = 10;
-            arm.setTargetPosition(liftTargetPosition);
-            arm.setPower(0.05);
-            arm.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-
-            sleep(500);
-            axle.setPosition(0.8);
-
-            while (arm.isBusy()) {}
-            arm.setPower(0);
 
     }
